requirements.txt
dearpygui==1.10.1
dearpygui_ext==0.9.5
einops==0.8.0
icecream==2.1.3
imageio==2.22.2
kornia==0.7.0
matplotlib==3.7.1
numpy==1.23.3
opencv_python==4.6.0.66
opencv_python_headless==4.9.0.80
Pillow==11.0.0
plyfile==1.1
pyexr==0.4.0
pyquaternion==0.9.9
scipy==1.10.0
setuptools==63.4.1
torch==2.0.1
torchvision==0.15.2 
tqdm==4.64.1
pyyaml>=6.0
trimesh>=4.0.0
openai>=1.0.0
requests>=2.31.0